% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{add_endpoint}
\alias{add_endpoint}
\title{add an endpoint to an URL}
\usage{
add_endpoint(url, endpoint)
}
\arguments{
\item{url}{a character}

\item{endpoint}{a character}
}
\description{
add an endpoint to an URL
}
\keyword{internal}
