% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{add_locale}
\alias{add_locale}
\title{add a locale to an URL}
\usage{
add_locale(url, locale)
}
\arguments{
\item{url}{a character}

\item{locale}{a character}
}
\description{
add a locale to an URL
}
\keyword{internal}
