% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{add_namespace}
\alias{add_namespace}
\title{add a namespace to an URL}
\usage{
add_namespace(args, namespace)
}
\arguments{
\item{namespace}{a character}

\item{url}{a character}
}
\description{
The Blizzard API uses namespaces for some endpoints, see
\href{https://develop.battle.net/documentation/guides/game-data-apis-wow-namespaces}{the doc}
for more context.
}
\keyword{internal}
