% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{augment_query}
\alias{augment_query}
\title{add default values to a request if needed}
\usage{
augment_query(...)
}
\arguments{
\item{...}{additional parameters to be passed to the request: token, region and/or locale}
}
\description{
add default values to a request if needed
}
