% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infer_locale}
\alias{infer_locale}
\title{check if a quantity has multiple locales}
\usage{
infer_locale(x, l)
}
\arguments{
\item{x}{an object, either a vector or a list}

\item{l}{a list, additional parameters, usually the \code{dots} variable}
}
\description{
Return a consistent structure whether it is single- or multi-locale.
}
