% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{make_get}
\alias{make_get}
\title{build a GET request}
\usage{
make_get(endpoint, token, region, namespace = NULL, locale = NULL)
}
\arguments{
\item{endpoint}{a character,}

\item{token}{a \code{\link[httr:oauth2.0_token]{httr::oauth2.0_token()}}}

\item{region}{a character, one of the available regions}

\item{namespace}{a character (optional)}

\item{locale}{a character (optional)}
}
\description{
build a GET request
}
\keyword{internal}
