% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{make_get_without_namespace}
\alias{make_get_without_namespace}
\title{build a GET request without a namespace}
\usage{
make_get_without_namespace(endpoint, ...)
}
\arguments{
\item{endpoint}{a character}

\item{...}{additional parameters to be passed to the request: token, region and/or locale}
}
\description{
build a GET request without a namespace
}
\keyword{internal}
