% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{region_base_url}
\alias{region_base_url}
\title{build the base URL for a region}
\usage{
region_base_url(region)
}
\arguments{
\item{region}{a character}
}
\description{
build the base URL for a region
}
\keyword{internal}
