% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{validate_token}
\alias{validate_token}
\title{validate a wowr token}
\usage{
validate_token(token)
}
\arguments{
\item{token}{an \code{\link[httr:oauth2.0_token]{httr::oauth2.0_token()}}}
}
\description{
The \code{validate()} function in \code{httr} has a hard-coded \emph{access_token} parameter to check for validity while the Blizzard API
uses \emph{token}, hence this custom function.
}
