% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{wow_auth}
\alias{wow_auth}
\title{authenticate to the Blizzard portal}
\usage{
wow_auth(client_id = NULL, client_secret = NULL, region = NULL)
}
\arguments{
\item{client_id}{a character, the application client id}

\item{client_secret}{a character, the application client secret id}

\item{region}{a character, the region to authenticate with}
}
\description{
authenticate to the Blizzard portal
}
