% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{wow_class}
\alias{wow_class}
\title{get detailed info on a specific class}
\usage{
wow_class(class_id, ...)
}
\arguments{
\item{class_id}{an integer, see \code{\link[=wow_classes]{wow_classes()}} for a full list}

\item{...}{additional parameters to be passed to the request: token, region and/or locale}
}
\description{
get detailed info on a specific class
}
