% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{wow_classes}
\alias{wow_classes}
\title{get all playable classes}
\usage{
wow_classes(...)
}
\arguments{
\item{...}{additional parameters to be passed to the request: token, region and/or locale}
}
\description{
get all playable classes
}
