% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realms.R
\name{wow_connected_realm}
\alias{wow_connected_realm}
\title{retrieve info on a specific connected realm}
\usage{
wow_connected_realm(realm_id, ...)
}
\arguments{
\item{realm_id}{an integer, see \code{\link[=wow_connected_realms]{wow_connected_realms()}} for a list}

\item{...}{additional parameters to be passed to the request: token, region and/or locale}
}
\description{
retrieve info on a specific connected realm
}
