% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realms.R
\name{wow_connected_realms}
\alias{wow_connected_realms}
\title{retrieve connected realms ids}
\usage{
wow_connected_realms(...)
}
\arguments{
\item{...}{additional parameters to be passed to the request: token, region and/or locale}
}
\description{
retrieve connected realms ids
}
