% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dungeons.R
\name{wow_dungeons}
\alias{wow_dungeons}
\title{get dungeons names and ids}
\usage{
wow_dungeons(...)
}
\arguments{
\item{...}{additional parameters to be passed to the request: token, region and/or locale}
}
\value{
a data.table with the following columns:
\itemize{
\item id,
\item name,
\item locale
}
}
\description{
get dungeons names and ids
}
