% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mythics.R
\name{wow_mythic_leaderboard}
\alias{wow_mythic_leaderboard}
\title{get the mythic leaderboard for a realm, dungeon and period}
\usage{
wow_mythic_leaderboard(realm_id, dungeon_id, period, ...)
}
\arguments{
\item{realm_id}{an integer, see \code{\link[=wow_realms]{wow_realms()}}}

\item{dungeon_id}{an integer, see \code{\link[=wow_dungeons]{wow_dungeons()}}}

\item{period}{an integer}

\item{...}{additional parameters to be passed to the request: token, region and/or locale}
}
\description{
get the mythic leaderboard for a realm, dungeon and period
}
