% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realms.R
\name{wow_realms}
\alias{wow_realms}
\title{retrieve realms ids}
\usage{
wow_realms(...)
}
\arguments{
\item{...}{additional parameters to be passed to the request: token, region and/or locale}
}
\description{
retrieve realms ids
}
