% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{wow_region}
\alias{wow_region}
\title{get detailed info on a specific region}
\usage{
wow_region(region_id, ...)
}
\arguments{
\item{region_id}{an integer, see \code{\link[=wow_regions]{wow_regions()}} for a full list}

\item{...}{additional parameters to be passed to the request: token, region and/or locale}
}
\description{
get detailed info on a specific region
}
