% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{wow_regions}
\alias{wow_regions}
\title{get all regions}
\usage{
wow_regions(...)
}
\arguments{
\item{...}{additional parameters to be passed to the request: token, region and/or locale}
}
\description{
get all regions
}
