% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{wow_specialization}
\alias{wow_specialization}
\title{get detailed info on a specific specialization}
\usage{
wow_specialization(spec_id, ...)
}
\arguments{
\item{spec_id}{an integer, see \code{\link[=wow_specializations]{wow_specializations()}} for a full list}

\item{...}{additional parameters to be passed to the request: token, region and/or locale}
}
\description{
get detailed info on a specific specialization
}
