% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{wow_specializations}
\alias{wow_specializations}
\title{get all playable specializations}
\usage{
wow_specializations(...)
}
\arguments{
\item{...}{additional parameters to be passed to the request: token, region and/or locale}
}
\description{
get all playable specializations
}
