% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wowr-package.R
\docType{package}
\name{wowr-package}
\alias{wowr}
\alias{wowr-package}
\title{wowr: An interface to the Wow API}
\description{
wowr provides functions to interact with the World of Warcraft API.
}
\author{
\strong{Maintainer}: Pierre Formont \email{pierre.formont@gmail.com}

}
\keyword{internal}
